Move AkkaHttpClient to new file, add AkkaHttpBackend.fromClient
